Extract Java cognitive complexity analysis
